Fix busy loop when connection problem
In case of connection problem, the collector threads might crash (for
example when wrong credentials provided). As a result, the processor
thread waits for content_types to become empty (which will never
happen). In the current implementation, the wait is implemented as a
busy loop.

This patch introduces a semaphor, so that the collector threads could
signal that they finished with their job. This also avoids busy wait.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>